Create rnn_models.model dynamically from model files
